Move API handler modules out of the Vercel api directory
The legacy `packages/app/api/` directory was left over from the
pre-Next.js layout. Its files are only meant to be imported by the App
Router route handlers, but Vercel's zero-config function detection also
deployed each one as a standalone serverless function, shadowing the
Next.js routes for `/api/*`.

Those functions run under Vercel's raw Node bootstrap, which hands the
handler a request whose `url` is path-only, so the OAuth callback threw
`ERR_INVALID_URL` from `new URL(req.url)` in production.

- Move ping, cdn (and its test), pr, and the GitHub OAuth callback to
  `src/server/`, matching the convention already used by
  `src/server/tokenAssets/`
- Rename `pr.tsx` to `pr.ts` since it contains no JSX
- Point the four `app/api/**/route.ts` handlers at the new paths

Next.js now serves `/api/*` exclusively, where `request.url` is always
absolute.
