+++ packages/app/src/server/cdn.test.ts
